:art: refactored: make GentlemenReportListener class-oriented listener
